pyproject.toml
[project]
name = "text-to-sql-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "fastapi>=0.127.0",
    "ipython-sql>=0.5.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.5",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "prettytable>=3.17.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "sqlglot>=28.5.0",
    "uvicorn>=0.40.0",
]
[tool.uv]
package = true
